Fix for [bug 2094] and [bug 2727].

+++ src/las.c
@@ -5442,7 +5442,7 @@ int dooptcode(struct pike_string *name,
 #endif
 	  ret=define_function(name,
 			      type,
-			      (unsigned INT8)modifiers,
+			      (unsigned INT16)modifiers,
 			      (unsigned INT8)(IDENTIFIER_C_FUNCTION | vargs),
 			      &tmp,
 			      foo->u.efun->flags);
@@ -5474,7 +5474,7 @@ int dooptcode(struct pike_string *name,
   
   ret=define_function(name,
 		      type,
-		      (unsigned INT8)modifiers,
+		      (unsigned INT16)modifiers,
 		      (unsigned INT8)(IDENTIFIER_PIKE_FUNCTION | vargs),
 		      Pike_compiler->num_parse_error?NULL:&tmp,
 		      (unsigned INT16)
